cloud_spanner/requirements.txt
google-cloud-spanner==1.3.0
futures==3.2.0; python_version < "3"
pandas==0.23.3